server/requirements.txt
django
djangorestframework
django_extensions
django-environ
psycopg2>=2.8
